Use Master label of Docker image (#16)
* Add Dockerfile and Compose file

Theoretically this is all you need for the bot to run in Docker but it might need some tweaking before it's ready. I'll test it later and add a GitHub Actions check

* Create .env.example

* Update executable path

* It works in Docker!

* Publish Docker image to ghcr on release

* Build image on push to master

* Add Action versions and dockerignore

GitHub didn't like not having versions for the Actions.
Also we probably don't want to have to copy our entire Git tree into the build context

* Create docker-compose.prod.yml

Theoretically compose tries to pull an image before it builds but that hasn't happened in my testing so we can't use an override file. Running in a production environment, use docker-compose.prod.yml

* Image must specify master branch