Leveling System (#3)
* Leveling helpers

* Implement RO redis interface

* Writeable Redis interface

* WIP: Message Handler

* Use new logging system

env_logger is not designed for use with async rust. `tracing` and
`tracing-subscriber` are, and are the recommended way for logging.

This is backwards-compatible with the `log` macros (e.g error! info!
warn! debug!), and should not require any further configuration.

All functions that use logging must have the `#[instrument]` macro, with
the exception of commands which automatically have that specified.

* WIP: Points given based on messages

Does not currently work - issue with writing to Redis silently.

* Write rank and levels commands

levels command is currently non-functional

* Remove unneeded debug prints

* Revert docker-compose.yml

* Remove redis dump

* Standardize redis `set` method