Ranked Leaderboard (#20)
* Leveling helpers

* Implement RO redis interface

* Writeable Redis interface

* WIP: Message Handler

* Use new logging system

env_logger is not designed for use with async rust. `tracing` and
`tracing-subscriber` are, and are the recommended way for logging.

This is backwards-compatible with the `log` macros (e.g error! info!
warn! debug!), and should not require any further configuration.

All functions that use logging must have the `#[instrument]` macro, with
the exception of commands which automatically have that specified.

* WIP: Points given based on messages

Does not currently work - issue with writing to Redis silently.

* Write rank and levels commands

levels command is currently non-functional

* Remove unneeded debug prints

* Revert docker-compose.yml

* Remove redis dump

* Standardize redis `set` method

* leveling: Make levels command use chunks to page

Previously used super jank for loops to page the leaderboard, now uses
chunks.

WIP, does not compile.

* leveling: use GuildId.members_iter to create leaderboard

Previously used only the users that appeared in the cache. Using the
members_iter iterator will fetch every single guild member.

Still will not sort the leaderboard properly -- I suspect that chunking
has something to do with it.

* util: Properly handle redis errors in set_user_level

Previously, set_user_level silently ignored errors. Now, it will print
error messages to the console and continue processing.

* chore: Run cargo fmt

* leveling: Make leaderboard sort after fetching from redis