Add clear, editmsg, and reactmasg (#33)
* staff: Add clear command

* staff: Add sendmsg command

* staff: Enable messages to be forwarded to staff

Adds a new event handler to reaction_add that will forward a copy of any
message that gains a ❗ or related reaction to a configurable staff
channel.

Configuration options added:
- LOGGING_CHANNEL (the ID of the channel to forward messages to)

* staff: Make message forwards support embeds

Will now resend the embed in the forward, allowing forwards from logs

* staff: Remove unneeded file downloading

Wasn't fully implemented (and won't be for now). Will create a new issue
once this is merged to work on it later.

* help: Add help command and document staff commands

* staff: Fix message forward channel mention

Now mentions the original channel, not the target channel

* staff: Add modmail

Modmail allows any member to DM the bot and have it sent to a special
channel. That channel will allow any mod to respond to the DM from the
bot.

* format

* staff: Add editmsg and reactmsg (WIP)

reactmsg is currently non-functional, but editmsg is.

Signed-off-by: Cara Salter <cara@devcara.com>

* staff: Finish reactmsg

Allows the bot to react to arbitrary messages

Signed-off-by: Cara Salter <cara@devcara.com>